feat: implemented Auth provider
Now the token is stored in local storage.
During login the JWT token is decoded, getting its id. The user object receives the username and id. Everything is stored in local storage.
The user provider received a function to make a request in the API and create a new habit.